feat: Separate access log filter into another package
- Support turn on/off request body and response body for full access log mode
- Support `operation` field in full access log mode, this is used to capture endpoint operation id
- Remove unused Legal filter

OS-6484
