+++ go.mod
@@ -6,12 +6,11 @@ require (
 	github.com/AccelByte/go-jose v2.1.4+incompatible
 	github.com/AccelByte/go-restful-plugins/v3 v3.2.2 // indirect
 	github.com/AccelByte/iam-go-sdk v1.7.1
-	github.com/AccelByte/legal-go-sdk v0.0.0-20210405152112-4c61182fa45b
 	github.com/AccelByte/public-source-ip v1.0.1
 	github.com/DataDog/datadog-go v4.3.0+incompatible // indirect
 	github.com/HdrHistogram/hdrhistogram-go v1.1.0 // indirect
 	github.com/emicklei/go-restful v2.15.0+incompatible // indirect
-	github.com/emicklei/go-restful/v3 v3.4.0
+	github.com/emicklei/go-restful/v3 v3.5.2
 	github.com/fatih/structs v1.1.0
 	github.com/google/uuid v1.1.5
 	github.com/json-iterator/go v1.1.10 // indirect
@@ -24,6 +23,7 @@ require (
 	github.com/uber/jaeger-lib v2.4.0+incompatible // indirect
 	github.com/willf/bitset v1.1.11 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
+	golang.org/x/crypto v0.0.0-20210322153248-0c34fe9e7dc2 // indirect
 	golang.org/x/sys v0.0.0-20210119212857-b64e53b001e4 // indirect
 	golang.org/x/time v0.0.0-20201208040808-7e3f01d25324 // indirect
 	gopkg.in/DataDog/dd-trace-go.v1 v1.28.0
