+++ pkg/logger/log/responsewriter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package common
+package log
 
 import "net/http"
